+++ src/components/TabMenu.js
@@ -5,7 +5,7 @@ const TabMenu = () => {
     return (
         <>
             <ul>
-                <li><Link to="/deposit" className={`${location.pathname == "/" ? "active" : location.pathname == "/deposit" ? "active" : ""}`}>Deposit</Link></li>
+                <li><Link to="/bridge/deposit" className={`${location.pathname == "/" ? "active" : location.pathname == "/bridge/deposit" ? "active" : ""}`}>Deposit</Link></li>
                 <li><Link to="/withdraw" className={`${location.pathname == "/withdraw" ? "active" : ""}`}>Withdraw</Link></li>
             </ul>
         </>
